Added the sequence calculations

+++ debt_snowball_calculator/internal/calculator/snowball.go
@@ -22,7 +22,7 @@ func (c *Snowball) Calculate(model *Model) DebtSequences {
 	rolloverPayment := 0.0
 	oneTimeImmediatePayment := model.Input.OneTimeImmediatePayment
 	compoundMinimumPayments := 0.0
-	maxMonth := 0
+	maxMonth := 0.0
 
 	sort.Slice(debts, func(i, j int) bool {
 		return debts[i].Amount < debts[j].Amount
@@ -35,12 +35,12 @@ func (c *Snowball) Calculate(model *Model) DebtSequences {
 
 		debtSequence := DebtSequence{
 			Debt:     debt,
-			Months:   []int{},
+			Months:   []float64{},
 			Payments: []float64{},
 			Balances: []float64{},
 		}
 
-		monthIter := 1
+		monthIter := 1.0
 		for {
 			basePayment := debt.MinimumPayment
